restore: split DDL tasks from collection task
